+++ db/migrate/20151022165943_create_committees.rb
@@ -1,0 +1,14 @@
+class CreateCommittees < ActiveRecord::Migration
+  def change
+    create_table :committees do |t|
+      t.string :name
+      t.string :slug
+      t.integer :position, null: false, default: 0
+      t.boolean :deleted, null: false, default: false
+
+      t.timestamps null: false
+    end
+
+    add_index :committees, :deleted
+  end
+end